refactor: try and further match existing DQS structure
