+++ src/client/pages/editorPage/CaseSummary.test.tsx
@@ -3,10 +3,10 @@ import {
 } from 'vitest';
 import { RenderResult, act, render } from '@testing-library/react';
 import { BrowserRouter } from 'react-router-dom';
-import { CaseSummaryDetails } from '../../../../common/interfaces/caseInterface';
+import { CaseSummaryDetails } from '../../../common/interfaces/caseInterface';
 import CaseSummary from './CaseSummary';
-import { getCaseSummary } from '../../../api/NodeApi';
-import { caseSummaryDetailsMockObject } from '../../../../server/test-utils/CaseMockObject';
+import { getCaseSummary } from '../../api/NodeApi';
+import { caseSummaryDetailsMockObject } from '../../../server/test-utils/CaseMockObject';
 
 vi.mock('react-router-dom', async () => {
   const actual = await vi.importActual('react-router-dom');
@@ -18,7 +18,7 @@ vi.mock('react-router-dom', async () => {
     }),
   };
 });
-vi.mock('../../../api/NodeApi');
+vi.mock('../../api/NodeApi');
 
 // declare global vars
 let view: RenderResult;
