+++ src/client/pages/editorPage/CaseSummary.tsx
@@ -1,11 +1,11 @@
 import { ReactElement } from 'react';
 import { useParams } from 'react-router-dom';
-import CaseSummaryContent from '../Components/CaseSummaryContent';
-import { CaseSummaryDetails } from '../../../../common/interfaces/caseInterface';
-import { CaseSummaryParams } from '../../../types/CaseSummaryParams';
-import AsyncContent from '../../shared/AsyncContent';
-import { useAsyncRequestWithTwoParams } from '../../../utils/useAsyncRequest';
-import { getCaseSummary } from '../../../api/NodeApi';
+import CaseSummaryContent from './sections/CaseSummaryContent';
+import { CaseSummaryDetails } from '../../../common/interfaces/caseInterface';
+import { CaseSummaryParams } from '../../types/CaseSummaryParams';
+import AsyncContent from '../shared/AsyncContent';
+import { useAsyncRequestWithTwoParams } from '../../utils/useAsyncRequest';
+import { getCaseSummary } from '../../api/NodeApi';
 
 function DisplayCaseSummary(questionnaireName: string, caseId: string) {
   const caseSummary = useAsyncRequestWithTwoParams<CaseSummaryDetails, string, string>(getCaseSummary, questionnaireName, caseId);
