+++ src/client/pages/editorPage/EditorHome.tsx
@@ -1,16 +1,16 @@
 import { User } from 'blaise-api-node-client';
-import { Survey } from '../../../../common/interfaces/surveyInterface';
-import { useAsyncRequestWithParam } from '../../../utils/useAsyncRequest';
-import { getSurveys } from '../../../api/NodeApi';
-import AsyncContent from '../../shared/AsyncContent';
-import SurveysList from '../../shared/SurveysList';
-import ErrorPanel from '../../shared/ErrorPanel';
+import AsyncContent from '../shared/AsyncContent';
+import { Survey } from '../../../common/interfaces/surveyInterface';
+import { getSurveys } from '../../api/NodeApi';
+import { useAsyncRequestWithParam } from '../../utils/useAsyncRequest';
+import SurveysList from '../shared/SurveysList';
+import ErrorPanel from '../shared/ErrorPanel';
 
 interface SurveyProps {
   user: User;
 }
 
-export default function SupportHome({ user }: SurveyProps) {
+export default function Surveys({ user }: SurveyProps) {
   const surveys = useAsyncRequestWithParam<Survey[], string>(getSurveys, user.role);
   const params = new URLSearchParams(window.location.search);
   const error = params.get('error') ?? '';
