+++ src/client/pages/editorPage/sections/CaseSummaryContent.tsx
