+++ src/client/pages/editorPage/sections/EditorContent.tsx
