+++ src/client/pages/researchPage/sections/ResearchContent.tsx
