+++ src/client/pages/supervisorPage/SupervisorHome.tsx
@@ -1,10 +1,10 @@
 import { User } from 'blaise-api-node-client';
-import { Survey } from '../../../../common/interfaces/surveyInterface';
-import { useAsyncRequestWithParam } from '../../../utils/useAsyncRequest';
-import { getSurveys } from '../../../api/NodeApi';
-import AsyncContent from '../../shared/AsyncContent';
-import SurveysList from '../../shared/SurveysList';
-import ErrorPanel from '../../shared/ErrorPanel';
+import { Survey } from '../../../common/interfaces/surveyInterface';
+import { useAsyncRequestWithParam } from '../../utils/useAsyncRequest';
+import { getSurveys } from '../../api/NodeApi';
+import AsyncContent from '../shared/AsyncContent';
+import SurveysList from '../shared/SurveysList';
+import ErrorPanel from '../shared/ErrorPanel';
 
 interface SurveyProps {
   user: User;
