+++ src/client/pages/supportPage/SupportHome.tsx
@@ -1,16 +1,16 @@
 import { User } from 'blaise-api-node-client';
-import AsyncContent from '../../shared/AsyncContent';
-import { Survey } from '../../../../common/interfaces/surveyInterface';
-import { getSurveys } from '../../../api/NodeApi';
-import { useAsyncRequestWithParam } from '../../../utils/useAsyncRequest';
-import SurveysList from '../../shared/SurveysList';
-import ErrorPanel from '../../shared/ErrorPanel';
+import { Survey } from '../../../common/interfaces/surveyInterface';
+import { useAsyncRequestWithParam } from '../../utils/useAsyncRequest';
+import { getSurveys } from '../../api/NodeApi';
+import AsyncContent from '../shared/AsyncContent';
+import SurveysList from '../shared/SurveysList';
+import ErrorPanel from '../shared/ErrorPanel';
 
 interface SurveyProps {
   user: User;
 }
 
-export default function Surveys({ user }: SurveyProps) {
+export default function SupportHome({ user }: SurveyProps) {
   const surveys = useAsyncRequestWithParam<Survey[], string>(getSurveys, user.role);
   const params = new URLSearchParams(window.location.search);
   const error = params.get('error') ?? '';
