+++ src/client/pages/supportPage/sections/SupportContent.tsx
